[MOV] im_livechat: rename WebsiteLivechat to PublicLivechat
Task-2894477
