+++ addons/im_livechat/static/src/legacy/widgets/livechat_button.js
@@ -9,7 +9,7 @@ import Widget from 'web.Widget';
 
 import { LIVECHAT_COOKIE_HISTORY } from 'im_livechat.legacy.im_livechat.Constants';
 import Feedback from '@im_livechat/legacy/widgets/feedback/feedback';
-import PublicLivechat from '@im_livechat/legacy/models/website_livechat';
+import PublicLivechat from '@im_livechat/legacy/models/public_livechat';
 import WebsiteLivechatMessage from '@im_livechat/legacy/models/website_livechat_message';
 import WebsiteLivechatWindow from '@im_livechat/legacy/models/website_livechat_window';
